Untangle LinqQueriesReadWrite test from LinqTestCase, moved domain classes and mappings for test into test assembly, thus fixed 'select from object' tests
