+++ src/NHibernate.Test/Linq/LinqTestCase.cs
@@ -33,8 +33,7 @@ namespace NHibernate.Test.Linq
 					"Northwind.Mappings.User.hbm.xml",
 					"Northwind.Mappings.TimeSheet.hbm.xml",
 					"Northwind.Mappings.Animal.hbm.xml",
-					"Northwind.Mappings.Patient.hbm.xml",
-					"ProductMaterial.hbm.xml"
+					"Northwind.Mappings.Patient.hbm.xml"
 				};
 			}
 		}
