+++ src/NHibernate.Test/Linq/ReadWrite/ProductMaterial.hbm.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<hibernate-mapping xmlns="urn:nhibernate-mapping-2.2" namespace="NHibernate.DomainModel" assembly="NHibernate.DomainModel" default-lazy="false">
+<hibernate-mapping xmlns="urn:nhibernate-mapping-2.2" namespace="NHibernate.Test.Linq.ReadWrite" assembly="NHibernate.Test" default-lazy="false">
 
     <class name="Material" >
         <id name="Id" type="Int32">
